🚚 rename: webp 확장자 변경
